backend_techinnoventia/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
djangorestframework==3.16.0
django-filter==24.3
djangorestframework_simplejwt==5.5.0
python-dotenv==1.0.1
djoser==2.3.1
idna==3.10
oauthlib==3.3.1
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
urllib3==2.5.0
